Restructure windows
