+++ src/BasicPanel.hpp
@@ -1,31 +1,26 @@
 #ifndef BASICPANEL_HPP
 #define BASICPANEL_HPP
 
+#include <map>
 #include "wx.hpp"
-
-class EnzymePanel;
-class ConfigurationPanel;
+#include "configuration.hpp"
 
 class BasicPanel : public wxPanel
 {
         private:
+                configuration &c;
+                std::map<wxString, enzyme> &enzymes;
                 wxListBox *fastaListBox;
                 wxListBox *xmlListBox;
-                wxCheckListBox *configurationCheckListBox;
-                ConfigurationPanel *configurationPanel;
                 wxCheckListBox *enzymeCheckListBox;
-                EnzymePanel *enzymePanel;
         public:
-                BasicPanel(wxWindow *parent, wxWindowID id);
-                void OnEnzDblClick(wxCommandEvent& event);
-                void OnConfDblClick(wxCommandEvent& event);
+                BasicPanel(wxWindow *parent, wxWindowID id, configuration &c_, std::map<wxString, enzyme> &enzymes_);
+                void update();
 };
 
 enum {
-        ID_XMLListBox = wxID_HIGHEST + 1,
-        ID_FastaListBox = wxID_HIGHEST + 2,
-        ID_ConfigurationCheckListBox = wxID_HIGHEST + 3,
-        ID_EnzymeCheckListBox = wxID_HIGHEST + 4
+        ID_FastaListBox = wxID_HIGHEST + 1,
+        ID_EnzymeCheckListBox = wxID_HIGHEST + 2
 };
 
 #endif
