+++ src/MainFrame.hpp
@@ -1,0 +1,34 @@
+#ifndef MAINFRAME_HPP
+#define MAINFRAME_HPP
+
+#include "wx.hpp"
+#include "ConfigurationPanel.hpp"
+#include "EnzymePanel.hpp"
+//#include <wx/notebook.h>
+//#include <wx/grid.h>
+
+class EnzymePanel;
+class ConfigurationPanel;
+
+class MainFrame : public wxFrame
+{
+        private:
+                wxCheckListBox *configurationCheckListBox;
+                ConfigurationPanel *configurationPanel;
+                wxListBox *enzymeListBox;
+                EnzymePanel *enzymePanel;
+                std::map<wxString, configuration> configurations;
+                std::map<wxString, enzyme> enzymes;
+        public:
+                MainFrame(const wxString& title);
+                void OnQuit(wxCommandEvent& event);
+                void OnEnzDblClick(wxCommandEvent& event);
+                void OnConfDblClick(wxCommandEvent& event);
+};
+
+enum {
+        wxID_ConfigurationCheckListBox = wxID_HIGHEST + 1,
+        wxID_EnzymeCheckListBox = wxID_HIGHEST + 2
+};
+
+#endif
